+++ app/src/main/java/com/jbrunoo/codinghealth/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.jbrunoo.codinghealth.local.TodoDatabase
+import com.jbrunoo.codinghealth.data.local.TodoDatabase
 import com.jbrunoo.codinghealth.ui.theme.CodingHealthTheme
 import com.jbrunoo.codinghealth.ui.theme.screen.TodoScreen
 import com.jbrunoo.codinghealth.vm.TodoViewModel
